requirements.txt
streamlit>=1.28.0
fastapi>=0.100.0
uvicorn>=0.20.0
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=40.0.0
Pillow>=10.0.0
PyPDF2>=3.0.0
plotly>=5.15.0
bcrypt>=4.0.0
pydantic>=2.0.0
email-validator>=2.0.0
python-dateutil>=2.8.0
requests>=2.30.0
aiofiles>=23.0.0
jinja2>=3.1.0
